% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pcoord.r
\name{add_pcoord}
\alias{add_pcoord}
\title{Principle Coordinates analysis of phylogenetic distance matrices}
\usage{
add_pcoord(D, correction = "none")
}
\arguments{
\item{D}{A list with objects H, P, and HP, as returned by \code{paco::prepare_paco_data}.}

\item{correction}{In some cases, phylogenetic distance matrices are non-Euclidean which generates negative eigenvalues when those matrices are translated into Principle Coordinates. There are several methods to correct negative eigenvalues. Correction options available here are "cailliez", "lingoes", and "none". The "cailliez" and "lingoes" corrections add a constant to the eigenvalues to make them non-negative. Default is "none".}
}
\value{
The list that was input as the argument `D' with two new elements; the Principle Coordinates of the `host' distance matrix and the Principle Coordinates of the `parasite' distance matrix.
}
\description{
Translates the distance matrices of 'host' and 'parasite' phylogenies into Principle Coordinates, as needed for Procrustes superimposition.
}
\note{
To find the Principle Coordinates of each distance matrix, we internally use a function, \code{coordpcoa}, that is a modified version of \code{ape::pcoa}, using \code{vegan::eigenvals}.
}
\examples{
data(gopherlice)
library(ape)
gdist <- cophenetic(gophertree)
ldist <- cophenetic(licetree)
D <- prepare_paco_data(gdist, ldist, gl_links)
D <- add_pcoord(D, correction='none')
}

