% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coordpcoa.r
\name{coordpcoa}
\alias{coordpcoa}
\title{Internal fucntion of add_pcoord performing priciple coordinate analysis}
\usage{
coordpcoa(D, correction = "none", rn = NULL)
}
\arguments{
\item{D}{A list with objects H, P, and HP, returned by prepare_paco_data}
}
\description{
Internal fucntion of add_pcoord performing priciple coordinate analysis
}
\note{
Internal function coordpcoa is a modified version of ape::pcoa, utilising vegan::eigenvals
}

