% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.npreg.R
\name{cycle.npreg.iter}
\alias{cycle.npreg.iter}
\title{Compute estimates in one iteration}
\usage{
\method{cycle}{npreg.iter}(Y, cell_times, ncores, ...)
}
\arguments{
\item{Y}{gene by sample log2CPM gene expression matrix}

\item{cell_times}{cell time estimate to be used in estimating cyclical patterns
of gene exprssion levels for each gene}

\item{ncores}{number of cores to be used}
}
