% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.trendfilter.generic.R
\name{fit.trendfilter.generic}
\alias{fit.trendfilter.generic}
\title{fitting trendfilter}
\usage{
fit.trendfilter.generic(yy, pos.yy = c(1:length(yy)), polyorder = 2)
}
\arguments{
\item{yy}{log2cpm gene expression vector}
}
