% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paco_data.r
\docType{data}
\name{gl_links}
\alias{gl_links}
\title{Gopher-lice interactions}
\usage{
data(gopherlice)
}
\description{
One part of example data. The associations between pocket gophers and their chewing lice ectoparasites.
}
\keyword{datasets}

