% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paco_data.r
\docType{data}
\name{gopherlice}
\alias{gopherlice}
\title{Pocket gopher and chewing lice example data}
\usage{
data(gopherlice)
}
\description{
Example dataset for cophylogeny. Includes a phylogeny for pocket gophers (gophertree), a phylogeny for their chewing lice parasites (licetree), and the interactions between them (gl_links).
}
\keyword{datasets}

