% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize a new RStudio project with a particular directory structure.}
\usage{
init(project_name, path = ".")
}
\arguments{
\item{project_name}{The name of the project (used in the README.txt and .Rproj file)}

\item{path}{The path where the project directory should be created}
}
\description{
Initialize a new RStudio project with a particular directory structure.
}
