% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PACo}
\alias{PACo}
\title{Performs PACo/procustes analysis}
\usage{
PACo(D, nperm = 1000, seed = NA, margin = 1)
}
\arguments{
\item{D}{a list with the data}

\item{nperm}{Number of permutations}

\item{seed}{Seed if results need to be reproduced}

\item{margin}{The margin to sample (1 to sample rows, 2 to sample columns)}
}
\description{
Performs PACo/procustes analysis
}
\examples{
data(gopherlice)
library(ape)
gdist <- cophenetic(gophertree)
ldist <- cophenetic(licetree)
D <- prepare_paco_data(gdist, ldist, gl_links)
D <- add_pcoord(D)
D <- PACo(D, nperm=10, seed=42)
print(D$gof)
}

