% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paco_links.r
\name{paco_links}
\alias{paco_links}
\title{Contribution of individual links}
\usage{
paco_links(D, ...)
}
\arguments{
\item{D}{A list returned by proc_analysis}

\item{...}{Additional arguments to be passed to PACo}
}
\value{
A list with added object jacknife, containing the mean and upper CI values for each link
}
\description{
Contribution of individual links
}

