% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pre_paco}
\alias{pre_paco}
\title{Prepare the data
Simple wrapper to make sure that the matrices are sorted accordingly}
\usage{
pre_paco(H, P, HP)
}
\arguments{
\item{H}{Host distance matrix}

\item{P}{Parasite distance matrix}

\item{HP}{Host-parasite association matrix, hosts in rows}
}
\value{
A list with objects H, P, HP
}
\description{
Prepare the data
Simple wrapper to make sure that the matrices are sorted accordingly
}

