% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prepare_paco_data}
\alias{prepare_paco_data}
\title{Prepare the data
Simple wrapper to make sure that the matrices are sorted accordingly}
\usage{
prepare_paco_data(H, P, HP)
}
\arguments{
\item{H}{Host distance matrix}

\item{P}{Parasite distance matrix}

\item{HP}{Host-parasite association matrix, hosts in rows}
}
\value{
A list with objects H, P, HP
}
\description{
Prepare the data
Simple wrapper to make sure that the matrices are sorted accordingly
}
\examples{
data(gopherlice)
library(ape)
gdist <- cophenetic(gophertree)
ldist <- cophenetic(licetree)
D <- prepare_paco_data(gdist, ldist, gl_links)
}

