% Generated by roxygen2 (4.0.2): do not edit by hand
\name{proc_analysis}
\alias{proc_analysis}
\title{Performs PACo/procustes analysis}
\usage{
proc_analysis(D, nperm = 1000, seed = NA)
}
\arguments{
\item{D}{a list with the data}

\item{nperm}{Number of permutations}

\item{seed}{Seed if results need to be reproduced}
}
\description{
Performs PACo/procustes analysis
}

