% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/residuals.paco.r
\name{residuals.paco}
\alias{residuals.paco}
\title{Get procrustes residuals from a paco object}
\usage{
\method{residuals}{paco}(object, ...)
}
\arguments{
\item{object}{a list with the data}

\item{...}{Used for type, wether the whole residual matrix (\code{matrix}) or the residuals per interaction (\code{interaction}) is desired}
}
\description{
Get procrustes residuals from a paco object
}

