% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/residuals_paco.r
\name{residuals_paco}
\alias{residuals_paco}
\title{Gets procrustes residuals from a paco object}
\usage{
residuals_paco(D, type = "interaction")
}
\arguments{
\item{D}{a list with the data}

\item{type}{Wether the whole residual matrix (\code{matrix}) or the residuals per interaction (\code{interaction}) is desired}
}
\description{
Gets procrustes residuals from a paco object
}

