% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_coverage.R
\name{summarize_coverage}
\alias{summarize_coverage}
\title{Summarize Coverage}
\usage{
summarize_coverage(df, ..., show_N = FALSE)
}
\arguments{
\item{df}{A data frame containing the columns \code{ISO_CODE}, \code{YEAR}, \code{COVERAGE_CODE}, \code{COVERAGE}, and \code{VALUE}.}

\item{...}{Additional grouping variables (unquoted) to group by.}

\item{show_N}{A logical value indicating whether to include the number of countries in each group in the result. Defaults to \code{FALSE}.}
}
\value{
A data frame with the calculated coverage for each group, including the total population for the group, and optionally the number of countries.
}
\description{
This function calculates the weighted coverage for a set of countries, using the population as the weights.
}
\details{
The function performs the following steps:
\itemize{
  \item Groups the data by \code{ISO_CODE}, \code{YEAR}, and \code{COVERAGE_CODE}.
  \item Filters out rows with missing \code{COVERAGE} or \code{VALUE}.
  \item Limits the \code{COVERAGE} to 100.
  \item Rounds the \code{COVERAGE} and \code{VALUE} columns.
  \item Optionally groups the data by additional columns provided in \code{...}.
  \item Calculates the total population for each group.
  \item Calculates the weighted coverage for each group.
  \item Summarizes the data to get the final coverage and population for each group.
  \item Optionally includes the number of countries in each group if \code{show_N} is \code{TRUE}.
}
}
\examples{
df <- data.frame(
  ISO_CODE = c("USA", "CAN", "MEX"),
  REGION = c("NOA", "NOA", "NOA"),
  YEAR = c(2020, 2020, 2020),
  COVERAGE_CODE = c("VAC1", "VAC1", "VAC1"),
  COVERAGE = c(95, 85, 90),
  VALUE = c(1000, 800, 1200)
)
summarize_coverage(df, show_N = FALSE)
# or with additional grouping
summarize_coverage(df, REGION, show_N = FALSE)

}
