% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations_to_list.R
\name{translations_to_list}
\alias{translations_to_list}
\title{Convert a translation data frame to a nested list}
\usage{
translations_to_list(
  data,
  language,
  id_module = "id_module",
  id_section = "id_section",
  id_element = "id_element"
)
}
\arguments{
\item{data}{A data frame containing translation data.}

\item{language}{A string specifying the name of the column containing translated values in a specific language.}

\item{id_module}{An optional string specifying the column to be used as the first level key of the list.}

\item{id_section}{An optional string specifying the column to be used as the second level key.}

\item{id_element}{An optional string specifying the column to be used as the third level key.}
}
\value{
A nested list with the structure module > section > element.
}
\description{
This function receives a data frame with PAHO-style translations and produces a nested list using specified ID columns and a language column.
}
\note{
Originally written by Alejandro Vasquez.
}
