% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_pcoord}
\alias{add_pcoord}
\title{paco}
\usage{
add_pcoord(D)
}
\arguments{
\item{D}{A list with objects H, P, and HP, returned by prepare_paco_data}
}
\value{
The input list with added objects for the principal coordinates of the objects
}
\description{
paco
}
\examples{
data(gopherlice)
library(ape)
gdist <- cophenetic(gophertree)
ldist <- cophenetic(licetree)
D <- prepare_paco_data(gdist, ldist, gl_links)
D <- add_pcoord(D)
print(D$gof)
}

