% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordpcoa.r
\name{coordpcoa}
\alias{coordpcoa}
\title{Internal fucntion of add_pcoord performing pricipal coordinate analysis}
\usage{
coordpcoa(D, correction = "none", rn = NULL)
}
\arguments{
\item{D}{A list with objects H, P, and HP, returned by prepare_paco_data}

\item{correction}{The correction to apply (none, lingoes, or cailliez)}

\item{rn}{rownames (optional)}
}
\description{
Internal fucntion of add_pcoord performing pricipal coordinate analysis
}
\note{
Internal function coordpcoa is a modified version of ape::pcoa, utilising vegan::eigenvals
}

