% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_from_WIISE.R
\name{fetch_from_WIISE}
\alias{fetch_from_WIISE}
\title{Fetch a Table from a public WIISEMart endpoint}
\usage{
fetch_from_WIISE(
  api,
  table_name,
  min_year = NULL,
  max_year = NULL,
  filters = list(),
  method = "get"
)
}
\arguments{
\item{api}{Character. A string specifying the base URL of the API endpoint.}

\item{table_name}{Character. A string specifying the name of the table to fetch from the API.}

\item{min_year}{Numeric (optional). A value specifying the minimum year of the data to fetch.}

\item{max_year}{Numeric (optional). A value specifying the maximum year of the data to fetch.}

\item{filters}{List (optional). A named list where each name is an additional column to filter by and each value is a character vector of values to include.}

\item{method}{Character (optional). Either "get" or "post". Defaults to "get" and automatically changes to "post" if URL size surpasses 2048 characters.}
}
\value{
A data frame containing the data from the specified table.
}
\description{
This function retrieves a specified table from WIISE and returns it as a data frame.
}
\examples{
\dontrun{
fetch_from_WIISE(
  api = "the_API_link",
  table_name = "your_table_name",
  min_year = 2022,
  max_year = 2023,
  filters = list(
    COUNTRY = c("COUNTRY_1", "COUNTRY_2"),
    INDICATOR = c("INDICATOR_1", "INDICATOR_2")
  ),
  method = "get"
)
}

}
