% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_from_WIISE.R
\name{fetch_from_WIISE}
\alias{fetch_from_WIISE}
\title{Fetch a Table from a public WIISEMart endpoint}
\usage{
fetch_from_WIISE(api, table_name, min_year = NULL, max_year = NULL)
}
\arguments{
\item{api}{A character string specifying the base URL of the API endpoint.}

\item{table_name}{A character string specifying the name of the table to fetch from the API.}

\item{min_year}{A numeric value specifying the minimum year of the data to fetch.}

\item{max_year}{A numeric value specifying the maximum year of the data to fetch.}
}
\value{
A data frame containing the data from the specified table.
}
\description{
This function retrieves a specified table from WIISE and returns it as a data frame.
}
