% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Create a directory structure according to PAHO/CIM standards.}
\usage{
init(project_name = "Insert project name here", path = ".")
}
\arguments{
\item{project_name}{The name of the project (used in the README.txt file). Defaults to "Insert project name here".}

\item{path}{The path where the project directory should be created. Defaults to current working directory.}
}
\description{
Create a directory structure according to PAHO/CIM standards.
}
