% Generated by roxygen2 (4.0.2): do not edit by hand
\name{paco_links}
\alias{paco_links}
\title{Contribution of individual links}
\usage{
paco_links(D)
}
\arguments{
\item{D}{A list returned by proc_analysis}
}
\value{
A list with added object jacknife, containing the mean and upper CI values for each link
}
\description{
Contribution of individual links
}

