% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paco_links.r
\name{paco_links}
\alias{paco_links}
\title{Contribution of individual links}
\usage{
paco_links(D, .parallel = FALSE, correction = "none")
}
\arguments{
\item{D}{A list returned by proc_analysis}

\item{.parallel}{if \code{TRUE}, calculate the jacknife contribution in parallel using the backend provided by foreach}

\item{correction}{Choose the method with which to correct negative eigenvalues in the internal call of add_pcoord ('none', cailliez', 'lingoes'). Default is 'none'.}
}
\value{
A list with added object jacknife, containing the mean and upper CI values for each link
}
\description{
Contribution of individual links
}

