% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_AMRO.R
\name{plot_map_AMRO}
\alias{plot_map_AMRO}
\title{Create a Map of the Americas with Caribbean Inset}
\usage{
plot_map_AMRO(
  shp,
  data,
  iso_col,
  fill_col,
  fill_levels = NULL,
  fill_labels = NULL,
  legend_title = "Value",
  palette = "PuBuGn"
)
}
\arguments{
\item{shp}{An \code{sf} object with country-level geometries and ISO3 codes.}

\item{data}{A data frame with a column matching \code{iso_col}, and a fill column (default: \code{DATA}).}

\item{iso_col}{Character. Name of the column with ISO3 codes in both \code{shp} and \code{data}.}

\item{fill_col}{Character. Name of the column with the data to use as fill.}

\item{fill_levels}{Character vector. Optional levels for the factor of \code{fill_col}.}

\item{fill_labels}{Character vector. Optional labels for factor levels of \code{fill_col}.}

\item{legend_title}{Character. Title for the legend. Default is "Value".}

\item{palette}{Character. Color palette to use in \code{scale_fill_brewer()}. Default is "PuBuGn".}
}
\value{
A \code{ggplot} object.
}
\description{
This function generates a map of the Americas (using provided ADM0 files) and paints it according
to the provide data. It includes a zoomed inset for the Caribbean.
}
